requirements.txt
django==5.0.6
djangorestframework==3.15.1
psycopg2-binary==2.9.9
celery==5.4.0
requests==2.32.1
drf-yasg==1.21.7
